Orientação a Objetos com Java/Semana4/Diferentes tipos de produto/produto/src/main/java/coursera/ita/Produto.java
package coursera.ita;

public class Produto {

	private String nome;
	protected String codigo;
	private double preco;

	public double getPreco() {
		return preco;
	}

	public String getNome() {
		return nome;
	}

	public Produto(String nome, String codigo, double preco) {
		super();
		this.nome = nome;
		this.codigo = codigo;
		this.preco = preco;
	}

	@Override
	public int hashCode() {
		final int prime = 31;
		int result = 1;
		result = prime * result + ((codigo == null) ? 0 : codigo.hashCode());
		return result;
	}

	@Override
	public boolean equals(Object obj) {
		if (this == obj)
			return true;
		if (obj == null)
			return false;
		if (getClass() != obj.getClass())
			return false;
		Produto other = (Produto) obj;
		if (codigo == null) {
			if (other.codigo != null)
				return false;
		} else if (!codigo.equals(other.codigo))
			return false;
		return true;
	}

}
